Fix signedness warning in gdk-pixbuf/gdk-pixbuf-data.c
gdk-pixbuf/gdk-pixbuf-data.c: In function ‘gdk_pixbuf_new_from_bytes’:
gdk-pixbuf/gdk-pixbuf-data.c:113:48: error: comparison of integer expressions of different signedness: ‘gsize’ {aka ‘long unsigned int’} and ‘int’ [-Werror=sign-compare]
  g_return_val_if_fail (g_bytes_get_size (data) >= width * height * (has_alpha ? 4 : 3), NULL);
                                                ^~
/usr/include/glib-2.0/glib/gmacros.h:427:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
gdk-pixbuf/gdk-pixbuf-data.c:113:2: note: in expansion of macro ‘g_return_val_if_fail’
  g_return_val_if_fail (g_bytes_get_size (data) >= width * height * (has_alpha ? 4 : 3), NULL);
  ^~~~~~~~~~~~~~~~~~~~

+++ gdk-pixbuf/gdk-pixbuf-data.c
@@ -110,7 +110,7 @@ gdk_pixbuf_new_from_bytes (GBytes *data, GdkColorspace colorspace, gboolean has_
 	g_return_val_if_fail (bits_per_sample == 8, NULL);
 	g_return_val_if_fail (width > 0, NULL);
 	g_return_val_if_fail (height > 0, NULL);
-	g_return_val_if_fail (g_bytes_get_size (data) >= width * height * (has_alpha ? 4 : 3), NULL);
+	g_return_val_if_fail (g_bytes_get_size (data) >= (gsize) (width * height * (has_alpha ? 4 : 3)), NULL);
 
 	return (GdkPixbuf*) g_object_new (GDK_TYPE_PIXBUF, 
 					  "pixel-bytes", data,
